Module onejoker::cards::card

wiki | A simple card object wrapping a u8.

Structs

Card

wiki | Card class A simple one-element tuple object wrapping an Ordinal value, which is just an alias for u8.

Constants

ACE_OF_CLUBS

ACE_OF_DIAMONDS

ACE_OF_HEARTS

ACE_OF_SPADES

BLACK_JOKER

DEUCE_OF_CLUBS

DEUCE_OF_DIAMONDS

DEUCE_OF_HEARTS

DEUCE_OF_SPADES

EIGHT_OF_CLUBS

EIGHT_OF_DIAMONDS

EIGHT_OF_HEARTS

EIGHT_OF_SPADES

FIVE_OF_CLUBS

FIVE_OF_DIAMONDS

FIVE_OF_HEARTS

FIVE_OF_SPADES

FOUR_OF_CLUBS

FOUR_OF_DIAMONDS

FOUR_OF_HEARTS

FOUR_OF_SPADES

JACK_OF_CLUBS

JACK_OF_DIAMONDS

JACK_OF_HEARTS

JACK_OF_SPADES

JOKER

KING_OF_CLUBS

KING_OF_DIAMONDS

KING_OF_HEARTS

KING_OF_SPADES

KNIGHT_OF_CLUBS

KNIGHT_OF_DIAMONDS

KNIGHT_OF_HEARTS

KNIGHT_OF_SPADES

LOW_ACE_OF_CLUBS

LOW_ACE_OF_DIAMONDS

LOW_ACE_OF_HEARTS

LOW_ACE_OF_SPADES

NINE_OF_CLUBS

NINE_OF_DIAMONDS

NINE_OF_HEARTS

NINE_OF_SPADES

QUEEN_OF_CLUBS

QUEEN_OF_DIAMONDS

QUEEN_OF_HEARTS

QUEEN_OF_SPADES

SEVEN_OF_CLUBS

SEVEN_OF_DIAMONDS

SEVEN_OF_HEARTS

SEVEN_OF_SPADES

SIX_OF_CLUBS

SIX_OF_DIAMONDS

SIX_OF_HEARTS

SIX_OF_SPADES

TEN_OF_CLUBS

TEN_OF_DIAMONDS

TEN_OF_HEARTS

TEN_OF_SPADES

TREY_OF_CLUBS

TREY_OF_DIAMONDS

TREY_OF_HEARTS

TREY_OF_SPADES

WHITE_JOKER

Type Aliases

Ordinal

wiki | Integer representation for cards Cards are represented as integers in the range 1..63