Module onejoker::cards::hand

wiki | Hand of cards

Structs

CardIntoIter

Iterator over Cards

CardIter

Iterator over Cards

Hand

wiki | Hand of cards A simple array of card objects with some utility methods. It is expected that most access will go through push/pop, which are fast, though things line insert and remove are available.