Struct onejoker::cards::hand::Hand

#[repr(C)]
pub struct Hand {
    pub cards: [Card; 22],
    pub length: u8,
    pub deck_type: u8,
}

wiki | Hand of cards A simple array of card objects with some utility methods. It is expected that most access will go through push()/pop(), which are fast, though things like insert() and remove() are available. Limited to 22 cards. If you need more, you can use Vec<Card>. but you lose some error checking and convenience methods.

Fields
cards: [Card; 22]

Array of Cards

length: u8

Number of cards in the hand

deck_type: u8

DeckType associated with this hand

Implementations

impl Hand

pub fn new(t: DeckType) -> Hand

Create new Hand associated with the given DeckType

use onejoker::*;
 
let h = Hand::new(DeckType::OneJoker);

pub fn new_by_name(dname: &str) -> Hand

Create new Hand associated with the DeckType of the given name

use onejoker::*;
 
let h = Hand::new_by_name("onejoker");

pub fn init<I>(self, iter: I) -> Self
where I: IntoIterator<Item = Card>,

Initialize new hand; return self for chaining

use onejoker::*;
 
let d = Deck::new(DeckType::English);
let h = d.new_hand().init(cards!("Qs", "Ac"));

pub fn sorted(self) -> Self

Initial sort for new hand; return self for chaining

use onejoker::*;
 
let d = Deck::new(DeckType::English);
let h = d.new_hand().init(cards!("Qs", "Ac")).sorted();
assert_eq!(h.to_string(), "AcQs");

pub fn deck_type(&self) -> DeckType

Return the DeckType associated with this hand

use onejoker::*;
 
let d = Deck::new_by_name("poker"); // alias for "english"
let h = d.new_hand();
assert_eq!(h.deck_type(), DeckType::English);

pub fn len(&self) -> usize

How many cards in the hand?

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let h = d.new_hand().init(d.draw_hand(cards!("Qs", "Ac")));
assert_eq!(h.len(), 2);

pub fn is_empty(&self) -> bool

Is the hand empty?

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let mut h = d.new_hand();
assert!(h.is_empty());
d.deal_to(&mut h, 5);
assert!(h.is_not_empty());

pub fn is_not_empty(&self) -> bool

Is the hand not empty?

pub fn clear(&mut self)

Empty the hand

use onejoker::*;
 
let mut h = Hand::default().init(cards!("As", "Qc"));
assert_eq!(h.len(), 2);
h.clear();
assert!(h.is_empty());

pub fn to_vec(&self) -> Vec<Card>

Export Vec of Cards (as copy)

use onejoker::*;
 
let h = Hand::default().init(cards!("As", "Qc"));
let v = h.to_vec();
assert_eq!(v.len(), 2);
assert_eq!(h.len(), 2);

pub fn as_slice(&self) -> &[Card]

Point to a slice of the underlying Card array

use onejoker::*;
 
let h = Hand::default().init(cards!("As", "Qc"));
let c: &[Card] = h.as_slice();

pub fn as_mut_slice(&mut self) -> &mut [Card]

Point to a mutable slice of the underlying Card array

use onejoker::*;
 
let mut h = Hand::default().init(cards!("As", "Qc"));
let c: &mut [Card] = h.as_mut_slice();

pub fn index_of(&self, card: Card) -> Option<usize>

Find given card in the hand, return index

use onejoker::*;
 
let h = Hand::default().init(cards!("Ac", "Kc", "Qc", "Jc", "Tc"));
assert_eq!(h.index_of(QUEEN_OF_CLUBS).unwrap(), 2);
assert_eq!(h.index_of(FOUR_OF_CLUBS), None);

pub fn contains(&self, card: Card) -> bool

Does the hand contain the given Card?

use onejoker::*;
 
let h = Hand::default().init(cards!("Ac", "Kc", "Qc", "Jc", "Tc"));
assert!(h.contains(QUEEN_OF_CLUBS));
assert!(! h.contains(FOUR_OF_CLUBS));

pub fn card_at(&self, index: usize) -> Option<Card>

Return the Card at the given index, or None if out of range

use onejoker::*;
 
let h = Hand::default().init(cards!("Ac", "Kc", "Qc", "Jc", "Tc"));
assert_eq!(h.card_at(2).unwrap(), QUEEN_OF_CLUBS);
assert_eq!(h.card_at(7), None);

pub fn set_card_at(&mut self, index: usize, card: Card) -> bool

Set the Card at the given index, or return false if out of range

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ac", "Kc", "Qc", "Jc", "Tc"));
assert_eq!(h.set_card_at(2, QUEEN_OF_DIAMONDS), true);
assert_eq!(h[2], QUEEN_OF_DIAMONDS);

pub fn push(&mut self, card: Card) -> bool

Push a Card onto the end of the hand

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh"));
h.push(QUEEN_OF_HEARTS);
assert_eq!(h.to_string(), "AhKhQh");

pub fn pop(&mut self) -> Option<Card>

Pop a Card from the end of the hand

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh"));
assert_eq!(h.pop().unwrap(), QUEEN_OF_HEARTS);
assert_eq!(h.to_string(), "AhKh");

pub fn push_n<I>(&mut self, n: usize, iter: I) -> usize
where I: IntoIterator<Item = Card>,

Push a collection of Cards onto the end of the hand, at most n, returning the number actually pushed.

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let mut h = d.new_hand().init(cards!("Ah", "Kh"));
assert_eq!(3, h.push_n(3, d.pop_n(3)));

pub fn push_all<I>(&mut self, iter: I) -> usize
where I: IntoIterator<Item = Card>,

Push a collection of Cards onto the end of the hand

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let mut h = d.new_hand().init(cards!("Ah", "Kh"));
h.push_all(d.pop_n(3));
assert_eq!(5, h.len());

pub fn pop_n(&mut self, n: usize) -> impl Iterator<Item = Card>

Pop n Cards from the end of the hand

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
let v: Vec<Card> = h.pop_n(3).collect();
assert_eq!(v.len(), 3);

pub fn pop_all(&mut self) -> impl Iterator<Item = Card>

Pop all Cards from the end of the hand

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
let v: Vec<Card> = h.pop_all().collect();
assert_eq!(v.len(), 5);

pub fn set<I>(&mut self, iter: I) -> bool
where I: IntoIterator<Item = Card>,

Replace the Cards in the hand with the given collection

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh"));
h.set(cards!("Qh", "Jh"));
assert_eq!(h.to_string(), "QhJh");

pub fn insert_at(&mut self, index: usize, card: Card) -> bool

Insert a Card at the given index

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Jh", "Th"));
h.insert_at(2, QUEEN_OF_HEARTS);
assert_eq!(h.to_string(), "AhKhQhJhTh");

pub fn remove_at(&mut self, index: usize) -> Option<Card>

Remove the Card at the given index

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
h.remove_at(2);
assert_eq!(h.to_string(), "AhKhJhTh");

pub fn remove_card(&mut self, card: Card) -> bool

Remove the given Card from the hand if present

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
h.remove_card(KING_OF_HEARTS);
assert_eq!(h.to_string(), "AhQhJhTh");

pub fn truncate(&mut self, n: usize)

Truncate the Hand to the given length

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
h.truncate(3);
assert_eq!(h.to_string(), "AhKhQh");

pub fn shuffle(&mut self)

Shuffle the Hand in place

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
h.shuffle();
println!("{}", h); // e.g., "QhJhKhThAh"

pub fn sort(&mut self)

Sort the Hand in place, descending by rank, then suit

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let mut h = d.new_hand().init(cards!("Th", "Kh", "Jh", "Ah", "Qh"));
h.sort();
assert_eq!(h.to_string(), "AhKhQhJhTh");

pub fn combinations(&self, k: usize) -> impl Iterator<Item = Hand>

Return an iterator over all n-card combinations of the hand

use onejoker::*;
 
let mut d = Deck::new(DeckType::English);
let h = d.new_hand().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
let mut count = 0;
for sub in h.combinations(3) {
    count += 1;
    println!("{}", sub);    // "AhKhQh", "AhKhJh", ...
}
assert_eq!(count, 10);

pub fn equals(&self, other: &Self) -> bool

Return true if the hands are identical: i.e., same cards in same order

use onejoker::*;
 
let h1 = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
let mut h2 = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
assert!(h1.equals(&h2));
h2.set(cards!("Th", "Ah", "Qh", "Jh", "Kh"));
assert!(! h1.equals(&h2));

pub fn is_equivalent_to(&self, other: &Self) -> bool

Return true if the hands are equivalent: i.e., same cards in any order

use onejoker::*;
 
let h1 = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
let h2 = Hand::default().init(cards!("Th", "Ah", "Qh", "Jh", "Kh"));
assert!(h1.is_equivalent_to(&h2));

pub fn ace_fix(&mut self)

Fix the ace values in the Hand to match the DeckType. Used internally–most users should not need this.

use onejoker::*;
 
let mut h = Hand::new(DeckType::Low).init(cards!("Ah", "2h"));
assert_eq!(h[0].rank(), Rank::LowAce);
// Unlike push() and set_card_at(), this does not error check
h[0] = ACE_OF_HEARTS;
assert_eq!(h[0].rank(), Rank::Ace);
h.ace_fix();
assert_eq!(h[0].rank(), Rank::LowAce);

pub fn discard(&mut self, indices: &[usize]) -> bool

Remove the cards at the given indices

use onejoker::*;
 
let mut h = Hand::default().init(cards!("Ah", "Kh", "Qh", "Jh", "Th"));
h.discard(&[1, 3]);
assert_eq!(h.to_string(), "AhQhTh");

impl Hand

pub fn iter(&self) -> CardIter

Return an iterator over the Cards in the hand.