Struct onejoker::poker::eval_state::EvaluatorState

pub struct EvaluatorState {
    pub cards: Vec<Card>,
    pub ranks: Vec<Rank>,
    pub wheel_is_straight: bool,
    pub sorted: Option<bool>,
    pub flush: Option<bool>,
    pub straight: Option<bool>,
    pub quads: Option<bool>,
    pub full_house: Option<bool>,
    pub trips: Option<bool>,
    pub two_pair: Option<bool>,
    pub pair: Option<bool>,
}

The straightforward way to evaluate poker hands requires making a lot of checks that are dependent on each other and their order of execution. This struct encapsulates the state of those checks and enforces the proper order. This code handles partial hands (fewer than 5 cards), but not hands > 5 cards.

Fields
cards: Vec<Card>

The cards in the hand.

ranks: Vec<Rank>

The ranks of the cards in the hand, maybe in different order

wheel_is_straight: bool

Does wheel count as a straight in this game?

sorted: Option<bool>

Are the ranks currently sorted (descending)?

flush: Option<bool>

Is the hand a flush?

straight: Option<bool>

Is the hand a straight?

quads: Option<bool>

Is the hand four of a kind?

full_house: Option<bool>

Is the hand a full house?

trips: Option<bool>

Is the hand three of a kind?

two_pair: Option<bool>

Is the hand two pair?

pair: Option<bool>

Is the hand one pair?

Implementations

impl EvaluatorState

pub fn new(hand: &Hand) -> Self

Create a new EvaluatorState object.

pub fn check_flush(&mut self) -> &mut Self

Check for flush

pub fn check_straight(&mut self) -> &mut Self

Check for straight

pub fn check_quads(&mut self) -> &mut Self

Check for four of a kind

pub fn check_full_house(&mut self) -> &mut Self

Check for full house

pub fn check_trips(&mut self) -> &mut Self

Check for three of a kind

pub fn check_two_pair(&mut self) -> &mut Self

Check for two pair

pub fn check_one_pair(&mut self) -> &mut Self

Check for one pair

pub fn all_checks_complete(&self) -> bool

Verify that all the checks have been made.

pub fn verify_no_pair(&self) -> bool

Verify that there are no pairs in the hand.