Module onejoker::poker::hand_evaluation

wiki | State machine for evaluating poker hands wiki | Traditional “high” poker hands

Structs

HandValue

Represent the complete result of evaluating a hand

Enums

HandLevel

wiki | Base class for categories of poker hands.

Functions

ojp_best_of

Given a large hand and evaluator, find the best 5-card hand.

ojp_best_value_of

Given a large hand and evaluator, find the best 5-card hand value.

ojp_default_eval_full

Default “full” hand evaluator for most poker games.

ojp_default_eval_quick

Default “value only” hand evaluator for most poker games.

ojp_default_hand_value

Hand value calculation that works for many high-hand games.

ojp_valid_hand_for_game

Is this a valid hand for the game?

Type Aliases

HandEvaluatorFull

Return full record of hand value info.

HandEvaluatorQuick

Return just numeric value to compare hands.