Module onejoker::poker::hand_value

wiki | Poker hand values base class wiki | Poker hand evaluation

Structs

HandValue

wiki | Hand value information structure All the information resulting for evaluating a poker hand.

Enums

HandLevelAceToFive

wiki | Ace-to-five lowball hand levels Ace-to-five lowball hands, e.g. Razz, California lowball, Stud/8, etc.

HandLevelAceToSix

wiki | Ace-to-six lowball hand levels Ace-to-six “London” lowball hands

HandLevelActionRazz

wiki | Action Razz hand levels Action Razz: any “qualified” hand with paint beats any “unqualified” hand.

HandLevelBadugi

wiki | Badugi hand levels Badugi: any four-card hand beats any 3-card, etc.

HandLevelDeuceToSeven

wiki | Deuce-to-seven lowball hand levels Deuce-to-seven lowball hands, e.g. Kansas City lowball, triple draw, etc.

HandLevelHigh

wiki | Traditional “high” poker hands

HandLevelStripped

wiki | Stripped-deck poker hands High poker hands for stripped deck games where flush beats full house.

Traits

HandEvaluatorTrait

wiki | Common layout of hand evaluators All specific game hand evaluators must implement this trait.

HandLevelTrait

wiki | Base class for categories of poker hands.

HandValueTrait

wiki | Hand value common code Common interface to hand value objects.

Functions

oj_default_ordered_for_display

Arrange the hand for display. E.g. “9d3h3cKs3s” -> “3s3h3cKs9d”.

oj_high_hand_value_function

Hand value calculation that works for many high-hand games.

oj_low_hand_value_function

Hand value calculation that works for many low-hand games.

Type Aliases

HandLevelManilla

wiki | Manilla poker hand levels Manilla uses stripped deck high hands.

HandLevelMexican

wiki | Mexican poker hand levels Mexican poker uses stripped deck high hands.

HandLevelPaiGow

wiki | Pai Gow hand levels Pai Gow uses same high hands, but slightly different rules.