Trait onejoker::poker::hand_value::HandEvaluatorTrait

pub trait HandEvaluatorTrait<V: HandValueTrait> {
    // Required method
    fn reference_evaluator(&self, _h: &Hand) -> aResult<V>;

    // Provided methods
    fn partial_evaluator(&self, h: &Hand) -> aResult<V> { ... }
    fn lookup_evaluator(&self, h: &Hand) -> aResult<V> { ... }
    fn fast_value(&self, h: &Hand) -> u64 { ... }
    fn value_of(&self, h: &Hand) -> aResult<V> { ... }
}

wiki | Common layout of hand evaluators All specific game hand evaluators must implement this trait.

Required Methods

fn reference_evaluator(&self, _h: &Hand) -> aResult<V>

Default evaluator, must implement.

Provided Methods

fn partial_evaluator(&self, h: &Hand) -> aResult<V>

Evaluator for partial hands, for determining stud betting, etc.

fn lookup_evaluator(&self, h: &Hand) -> aResult<V>

Fast lookup-table based evaluator.

fn fast_value(&self, h: &Hand) -> u64

Fast no-frills evaluator if all you care about is what wins.

fn value_of(&self, h: &Hand) -> aResult<V>

General-use entry point that picks the right evaluator.