Trait onejoker::poker::hand_value::HandEvaluatorTrait

pub trait HandEvaluatorTrait<V: HandValueTrait> {
    const COMPLETE_HAND: usize = 5usize;

    // Required method
    fn reference_evaluator(&self, _h: &Hand) -> Result<V, OjError>;

    // Provided methods
    fn partial_evaluator(&self, h: &Hand) -> Result<V, OjError> { ... }
    fn lookup_evaluator(&self, h: &Hand) -> Result<V, OjError> { ... }
    fn fast_value(&self, h: &Hand) -> u64 { ... }
    fn value_of(&self, h: &Hand) -> Result<V, OjError> { ... }
}

wiki | Common layout of hand evaluators All specific game hand evaluators must implement this trait.

Provided Associated Constants

const COMPLETE_HAND: usize = 5usize

Number of cards in a complete hand.

Required Methods

fn reference_evaluator(&self, _h: &Hand) -> Result<V, OjError>

Default evaluator, must implement.

Provided Methods

fn partial_evaluator(&self, h: &Hand) -> Result<V, OjError>

Evaluator for partial hands, for determining stud betting, etc.

fn lookup_evaluator(&self, h: &Hand) -> Result<V, OjError>

Fast lookup-table based evaluator.

fn fast_value(&self, h: &Hand) -> u64

Fast no-frills evaluator if all you care about is what wins.

fn value_of(&self, h: &Hand) -> Result<V, OjError>

General-use entry point that picks the right evaluator.

Object Safety

This trait is not object safe.